refactor: 데이터 모델별로 파일을 분리
